Split stateful.rs into more modules
